Combined Plugins Tests refactoring
